fix(settings): a mint in flight cannot be closed, and its tests move out
Three findings from review, all of them right.

THE ONE THAT MATTERS IS NOT LAYOUT. `mint.reset()` detaches the observer; it
does not cancel the POST. So a drawer closed mid-flight still created a passport
on the server — and its token, disclosed exactly once and never re-served, went
with the drawer. The list carries metadata only, so there was no way back: an
orphaned credential nobody holds. `closeMint` refuses while the request is
outstanding, and Cancel is disabled with it. The test hangs the request rather
than racing a window, and was checked against the unfixed code to confirm it
actually catches this.

THE MINT TESTS MOVE TO THEIR OWN FILE. `settings.test.tsx` is already well past
the 1000-line ceiling `frontend/CLAUDE.md` sets, and this had added 157 more to
it. A suite stops being navigable long before it stops running.

AND THEY DRIVE ONE `userEvent` INSTANCE PER TEST rather than calling the module
import per interaction, which is the same file's rule: a second instance
silently forgets which keys and buttons the first left held.
