+++ frontend/src/screens/settings.test.tsx
@@ -682,164 +682,6 @@ describe("PassportCard revoke (AS-2)", () => {
   });
 });
 
-// The mint flow's own backend: the POST answers with a token exactly once, and
-// the caller decides whether it succeeds, so a refusal can be exercised without
-// a second fixture.
-function mintBackend(opts: { refuse?: boolean } = {}) {
-  return vi.fn(async (input: RequestInfo | URL) => {
-    const url = String(input instanceof Request ? input.url : input);
-    const method = input instanceof Request ? input.method : "GET";
-    if (url.endsWith("/v1/me")) {
-      return jsonResponse({
-        user: { email: "ada@acme.test" },
-        roles: ["admin"],
-        teams: [],
-      });
-    }
-    if (url.includes("/passports") && method === "POST") {
-      if (opts.refuse) {
-        return jsonResponse(
-          {
-            type: "about:blank",
-            title: "Forbidden",
-            status: 403,
-            detail: "Your seat cannot lend that scope.",
-          },
-          403,
-        );
-      }
-      return jsonResponse({
-        id: "pp-new",
-        label: "Scout",
-        scopes: ["read", "draft"],
-        created_at: "2026-08-01T08:00:00Z",
-        expires_at: null,
-        revoked_at: null,
-        token: "mgp_live_0f3a91c4",
-      });
-    }
-    if (url.includes("/passports")) {
-      return jsonResponse({
-        data: [],
-        page: { next_cursor: null, has_more: false },
-      });
-    }
-    return jsonResponse({
-      data: [],
-      page: { next_cursor: null, has_more: false },
-    });
-  });
-}
-
-// Minting is a credential-issuing flow whose result is shown exactly once, and
-// it had no test at all while it was a row of controls inside the card. These
-// hold what the drawer owes a reader: a named group of choices, a refusal it
-// cannot submit past, the token surfaced where it cannot be missed, and a
-// dialog that does not take that token away by closing itself.
-describe("PassportCard — minting", () => {
-  async function openDrawer() {
-    render(<SettingsScreen tab="agents" />);
-    const open = await screen.findByRole("button", { name: "Mint passport" });
-    await userEvent.click(open);
-    return screen.findByRole("dialog");
-  }
-
-  it("puts the form in a drawer, with the scopes as a named group", async () => {
-    vi.stubGlobal("fetch", mintBackend());
-    const dialog = await openDrawer();
-
-    // The name field is a real label, not a span pointed at by aria-labelledby:
-    // clicking the words has to focus the control.
-    const name = within(dialog).getByLabelText("Agent name");
-    await userEvent.click(within(dialog).getByText("Agent name"));
-    expect(name).toHaveFocus();
-
-    // Five choices that belong together ARE a group, and the group has a name.
-    const group = within(dialog).getByRole("group", {
-      name: /what this agent may do/i,
-    });
-    expect(within(group).getAllByRole("checkbox")).toHaveLength(5);
-  });
-
-  it("refuses a passport with no scope, and says why", async () => {
-    vi.stubGlobal("fetch", mintBackend());
-    const dialog = await openDrawer();
-    const group = within(dialog).getByRole("group", {
-      name: /what this agent may do/i,
-    });
-
-    // read and draft are on by default; clearing both leaves a credential that
-    // could do nothing.
-    for (const box of within(group).getAllByRole("checkbox")) {
-      if ((box as HTMLInputElement).checked) await userEvent.click(box);
-    }
-
-    const submit = within(dialog).getByRole("button", {
-      name: "Mint passport",
-    });
-    expect(submit).toBeDisabled();
-    // Refused, and the reason is attached to the control rather than left in a
-    // title no screen reader announces on a disabled button.
-    const describedBy = submit.getAttribute("aria-describedby");
-    expect(describedBy).toBeTruthy();
-    expect(document.getElementById(describedBy ?? "")?.textContent).toMatch(
-      /at least one/i,
-    );
-  });
-
-  it("shows the token once, moves focus to it, and does not close itself", async () => {
-    vi.stubGlobal("fetch", mintBackend());
-    const dialog = await openDrawer();
-    await userEvent.type(within(dialog).getByLabelText("Agent name"), "Scout");
-    await userEvent.click(
-      within(dialog).getByRole("button", { name: "Mint passport" }),
-    );
-
-    const token = await within(dialog).findByText("mgp_live_0f3a91c4");
-    expect(token).toBeTruthy();
-    // The region is a live one and focus lands in it: the token is disclosed
-    // exactly once, so a reader whose focus stayed on the button would have to
-    // hunt for what they just made.
-    const region = token.closest('[role="status"]');
-    expect(region).toBeTruthy();
-    expect(region).toHaveFocus();
-    // Still open. Closing on success would take the only sight of the
-    // credential with it.
-    expect(screen.getByRole("dialog")).toBeTruthy();
-  });
-
-  it("announces a refused mint beside the button that produced it", async () => {
-    vi.stubGlobal("fetch", mintBackend({ refuse: true }));
-    const dialog = await openDrawer();
-    await userEvent.click(
-      within(dialog).getByRole("button", { name: "Mint passport" }),
-    );
-
-    const alert = await within(dialog).findByRole("alert");
-    expect(alert).toHaveTextContent(/cannot lend that scope/i);
-    // In the form, not in a band elsewhere on the card.
-    expect(alert.closest("form")).toBeTruthy();
-  });
-
-  it("starts clean on re-open rather than showing the last mint's token", async () => {
-    vi.stubGlobal("fetch", mintBackend());
-    const dialog = await openDrawer();
-    await userEvent.click(
-      within(dialog).getByRole("button", { name: "Mint passport" }),
-    );
-    await within(dialog).findByText("mgp_live_0f3a91c4");
-    await userEvent.click(within(dialog).getByRole("button", { name: "Done" }));
-
-    await waitFor(() => expect(screen.queryByRole("dialog")).toBeNull());
-    await userEvent.click(
-      screen.getByRole("button", { name: "Mint passport" }),
-    );
-    const reopened = await screen.findByRole("dialog");
-    expect(within(reopened).queryByText("mgp_live_0f3a91c4")).toBeNull();
-    expect(within(reopened).getByLabelText("Agent name")).toHaveValue("");
-  });
-});
-
 describe("SettingsScreen tab layout", () => {
   // These layout assertions run as an admin holding the org grants, so every
   // tab under test is present. Which principal sees which tab is the
